+++ Sources/TuistCore/Extensions/ArgumentParserError+FatalError.swift
@@ -1,0 +1,10 @@
+import Utility
+
+// MARK: - FatalError
+
+extension ArgumentParserError: FatalError {
+    /// ArgumentParserError has the .abort error type by default.
+    public var type: ErrorType {
+        return .abort
+    }
+}