+++ Sources/TuistKit/Embed/Embeddable.swift
@@ -91,7 +91,7 @@ final class Embeddable {
 
     func architectures(system: Systeming = System()) throws -> [String] {
         guard let binPath = try binaryPath() else { return [] }
-        let lipoResult = try system.capture("lipo", "-info", binPath.asString, verbose: false).throwIfError().stdout.chuzzle() ?? ""
+        let lipoResult = try system.capture("/usr/bin/lipo", arguments: "-info", binPath.asString, verbose: false, environment: nil).stdout.chuzzle() ?? ""
         var characterSet = CharacterSet.alphanumerics
         characterSet.insert(charactersIn: " _-")
         let scanner = Scanner(string: lipoResult)
@@ -170,7 +170,10 @@ final class Embeddable {
     fileprivate func stripArchitecture(packagePath: AbsolutePath,
                                        architecture: String,
                                        system: Systeming = System()) throws {
-        try system.popen("lipo", "-remove", architecture, "-output", packagePath.asString, packagePath.asString, verbose: false)
+        try system.popen("/usr/bin/lipo",
+                         arguments: "-remove", architecture, "-output", packagePath.asString, packagePath.asString,
+                         verbose: false,
+                         environment: nil)
     }
 
     fileprivate func stripHeaders(frameworkPath: AbsolutePath) throws {
@@ -217,7 +220,10 @@ final class Embeddable {
 
     fileprivate func uuidsFromDwarfdump(path: AbsolutePath,
                                         system: Systeming = System()) throws -> Set<UUID> {
-        let result = try system.capture("dwarfdump", "--uuid", path.asString, verbose: false).throwIfError().stdout.chuzzle() ?? ""
+        let result = try system.capture("/usr/bin/dwarfdump",
+                                        arguments: "--uuid", path.asString,
+                                        verbose: false,
+                                        environment: nil).stdout.chuzzle() ?? ""
         var uuidCharacterSet = CharacterSet()
         uuidCharacterSet.formUnion(.letters)
         uuidCharacterSet.formUnion(.decimalDigits)