+++ Tests/TuistEnvKitTests/Installer/InstallerTests.swift
@@ -73,7 +73,7 @@ final class InstallerTests: XCTestCase {
             .path
             .appending(component: Constants.bundleName)
         system.stub(args: [
-            "curl", "-LSs",
+            "/usr/bin/curl", "-LSs",
             "--output", downloadPath.asString,
             downloadURL.absoluteString,
         ],
@@ -81,7 +81,7 @@ final class InstallerTests: XCTestCase {
                     stdout: nil,
                     exitstatus: 0)
         system.stub(args: [
-            "unzip", downloadPath.asString,
+            "/usr/bin/unzip", downloadPath.asString,
             "-d", self.fileHandler.currentPath.asString,
         ],
                     stderror: nil,
@@ -120,7 +120,7 @@ final class InstallerTests: XCTestCase {
             .path
             .appending(component: Constants.bundleName)
         system.stub(args: [
-            "curl", "-LSs",
+            "/usr/bin/curl", "-LSs",
             "--output", downloadPath.asString,
             downloadURL.absoluteString,
         ],
@@ -155,7 +155,7 @@ final class InstallerTests: XCTestCase {
             .path
             .appending(component: Constants.bundleName)
         system.stub(args: [
-            "curl", "-LSs",
+            "/usr/bin/curl", "-LSs",
             "--output", downloadPath.asString,
             downloadURL.absoluteString,
         ],
@@ -163,7 +163,7 @@ final class InstallerTests: XCTestCase {
                     stdout: nil,
                     exitstatus: 0)
         system.stub(args: [
-            "unzip", downloadPath.asString,
+            "/usr/bin/unzip", downloadPath.asString,
             "-d", self.fileHandler.currentPath.asString,
         ],
                     stderror: "unzip_error",
@@ -186,7 +186,7 @@ final class InstallerTests: XCTestCase {
         }
 
         system.stub(args: [
-            "git",
+            "/usr/bin/env", "git",
             "clone", Constants.gitRepositoryURL,
             temporaryDirectory.path.asString,
         ],
@@ -194,7 +194,7 @@ final class InstallerTests: XCTestCase {
                     stdout: nil,
                     exitstatus: 0)
         system.stub(args: [
-            "git", "-C", temporaryDirectory.path.asString,
+            "/usr/bin/env", "git", "-C", temporaryDirectory.path.asString,
             "checkout", version,
         ],
                     stderror: nil,
@@ -254,7 +254,7 @@ final class InstallerTests: XCTestCase {
         }
 
         system.stub(args: [
-            "git",
+            "/usr/bin/env", "git",
             "clone", Constants.gitRepositoryURL,
             temporaryDirectory.path.asString,
         ],
@@ -262,7 +262,7 @@ final class InstallerTests: XCTestCase {
                     stdout: nil,
                     exitstatus: 0)
         system.stub(args: [
-            "git", "-C", temporaryDirectory.path.asString,
+            "/usr/bin/env", "git", "-C", temporaryDirectory.path.asString,
             "checkout", version,
         ],
                     stderror: "did not match any file(s) known to git ",